fix(core): githooks are initialized in wrong directory
