+++ renku/command/githooks.py
@@ -18,14 +18,14 @@
 """Install and uninstall Git hooks."""
 
 from renku.command.command_builder.command import Command
-from renku.core.githooks import install, uninstall
+from renku.core.githooks import install_githooks, uninstall_githooks
 
 
 def install_githooks_command():
     """Command to install Git hooks."""
-    return Command().command(install)
+    return Command().command(install_githooks)
 
 
 def uninstall_githooks_command():
     """Command to uninstall Git hooks."""
-    return Command().command(uninstall)
+    return Command().command(uninstall_githooks)
