+++ tests/cli/test_init.py
@@ -81,6 +81,7 @@ def test_init(isolated_runner, project_init):
     result = isolated_runner.invoke(cli, commands["init_test"] + commands["id"], commands["confirm"])
     assert 0 == result.exit_code, format_result_exception(result)
     assert new_project.exists()
+    assert "RENKU HOOK. DO NOT REMOVE OR MODIFY." in (new_project / ".git" / "hooks" / "pre-commit").read_text()
     assert (new_project / ".renku").exists()
     assert (new_project / ".renku" / "renku.ini").exists()
     assert (new_project / ".renku" / "metadata").exists()
